+++ src/main/java/cart/dao/item/ItemDao.java
@@ -1,4 +1,4 @@
-package cart.dao;
+package cart.dao.item;
 
 
 import cart.entity.ItemEntity;
