+++ src/main/java/cart/dao/member/MemberDao.java
@@ -1,4 +1,4 @@
-package cart.dao;
+package cart.dao.member;
 
 import cart.entity.MemberEntity;
