+++ src/test/java/cart/dao/JdbcItemDaoTest.java
@@ -1,5 +1,6 @@
 package cart.dao;
 
+import cart.dao.item.ItemDao;
 import cart.entity.ItemEntity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
